(UI) Adding a clean homescreen with just open repo option
