+++ src/components/AppInner.tsx
@@ -1,0 +1,13 @@
+import { Component } from "solid-js";
+
+import { useRepository } from "../stores/repository";
+import Home from "./Home";
+import Player from "./Player";
+
+const AppInner: Component = () => {
+  const [store] = useRepository();
+
+  return <>{store.isReady ? <Player /> : <Home />}</>;
+};
+
+export default AppInner;
