+++ src/components/Controls.tsx
@@ -2,22 +2,7 @@ import { Component, createEffect, createSignal, onMount } from "solid-js";
 import type { JSX } from "solid-js";
 
 import { useRepository } from "../stores/repository";
-
-interface IButtonPropTypes {
-  label: string;
-  onClick?: JSX.EventHandler<HTMLButtonElement, MouseEvent>;
-}
-
-const Button: Component<IButtonPropTypes> = (props: IButtonPropTypes) => {
-  return (
-    <button
-      class="p-1 px-4 mx-2 text-sm rounded-md border-gray-300 border bg-white hover:shadow-sm hover:bg-gray-500 hover:text-white"
-      onClick={props.onClick}
-    >
-      {props.label}
-    </button>
-  );
-};
+import Button from "./Button";
 
 interface IRepositoryFormPropTypes {
   setToggle: (value: boolean) => void;
